flyio_get_dir added in import_table

+++ R/import_table.R
@@ -1,9 +1,12 @@
 #' Read csv, Excel files, txt
 #' @description Read tabular data from anywhere using a function defined by you
+#'
 #' @param file path of the file to be read
 #' @param FUN the function using which the file is to be read
 #' @param data_source the name of the data source, if not set globally. s3, gcs or local
 #' @param bucket the name of the bucket, if not set globally
+#' @param dir the directory to store intermediate files
+#' @param delete_file logical. to delete the file downloaded
 #' @param ... other parameters for the FUN function defined above
 #'
 #' @export "import_table"
@@ -17,7 +20,7 @@
 #' }
 
 import_table <- function(file, FUN = read.csv, data_source = flyio_get_datasource(),
-                         bucket = flyio_get_bucket(data_source), ...){
+                         bucket = flyio_get_bucket(data_source), dir = flyio_get_dir(), delete_file = TRUE, ...){
   # checking if the file is valid
   assert_that(tools::file_ext(file) %in% c("csv", "xlsx", "xls", "txt"), msg = "Please input a valid path")
   if(data_source == "local"){
@@ -25,8 +28,8 @@ import_table <- function(file, FUN = read.csv, data_source = flyio_get_datasourc
     return(t)
   }
   # a tempfile with the required extension
-  temp <- tempfile(fileext = paste0(".",tools::file_ext(file)))
-  on.exit(unlink(temp))
+  temp <- paste0(dir, "/", basename(file))
+  if(isTRUE(delete_file)){on.exit(unlink(temp))}
   # downloading the file
   file = gsub("\\/+","/",file)
   downlogical = import_file(bucketpath = file, localfile = temp, data_source = data_source, bucket = bucket)
